AG-12512 Check InteractionState in KeyNavManager.addListener
Previously, this check was done by the RegionManager class. But now, the
RegionManager only handles pointer events. So restore this check.

This fixes a regression with an e2e test with financial-charts-test.
